requirements.txt
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
sqlalchemy
python-dotenv
fastapi 
nest-asyncio 
pyngrok 
uvicorn
streamlit